requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fsspec==2024.12.0
google-auth==2.39.0
google-genai==1.11.0
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.55
langchain-neo4j==0.4.0
langchain-text-splitters==0.3.8
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.33
mem0ai==0.1.93
monotonic==1.6
mypy_extensions==1.1.0
neo4j==5.28.1
neo4j-graphrag==1.6.1
numpy==2.2.5
ollama==0.4.8
openai==1.75.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
portalocker==2.10.1
posthog==3.25.0
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.14.1
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250402
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
xxhash==3.5.0
zstandard==0.23.0
